test(accordion): migrate a11y test
